dom/tests/mochitest/mouselock/test_DOMtree.html

<!DOCTYPE HTML>
<html>
    <!--
    https://bugzilla.mozilla.org/show_bug.cgi?id=633602
    -->
    <head>
        <title>Test for Bug 633602</title>
        <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
        <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
        <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
    </head>
    <body>
        <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">Mozilla Bug 633602</a>
        <div id="content">
        </div>
        <pre id="test">
            <script type="text/javascript">
                /** Test for Bug 633602 **/
                SimpleTest.waitForExplicitFinish()
                SimpleTest.waitForFocus(function() {
                    /*------------test start----------*/
                    
                    SpecialPowers.setBoolPref("full-screen-api.allow-trusted-requests-only", false);
                    
                    var pointer = navigator.pointer;
                    var canvas = document.createElement('canvas');
                    canvas.id = "testCanvas";
                    canvas.width = 250;
                    canvas.height = 250;
                    canvas.className = "test";
                    
                    pointer.lock(canvas);
                    is(pointer.islocked(), false, "Mouse should NOT be locked when DOM element is not in the tree.");
                    /*if (!pointer.islocked()) {
                        console.log("mouse lock should not happen");
                    }*/
                    
                    
                    var bodyTag = document.getElementsByTagName('body')[0];
                    bodyTag.appendChild(canvas);
                    //pointer.lock(canvas);
                    //runTests();
                    document.body.mozRequestFullScreen();
                    //document.body.click();
                    /*is(pointer.islocked(), true, "Mouse should sdfdsfbe locked when DOM element is in the tree."); //if still fail, throw to inside the mozfullscreenchange event
                    /*if (pointer.islocked()) {
                        console.log("mouse lock should happen");
                    }
                    document.mozCancelFullScreen();*/
                    
                    bodyTag.removeChild(canvas);
                    is(pointer.islocked(), false, "Mouse should NOT be locked when DOM element is removed from the tree.");
                    /*if (!pointer.islocked()) {
                        console.log("Mouse should NOT be locked when DOM element is removed from the tree");
                    }*/
                    /*
                    if (pointer.islocked()) {
                        console.log("=== Mouse lock enabled but should not happen! ===");
                        setTimeout("removeElement()", 2000);
                    }*/
                    
                    /*var bodyTag = document.getElementsByTagName('body')[0];
                    bodyTag.appendChild(canvas);
                    /*bodyTag.onclick = function () {
			canvas.mozRequestFullScreen();
                        canvas.__curX = 0;
			canvas.__curY = 0;
                    };*/

                    //
                    
                    /*document.body.addEventListener('click', function() {
				if (!document.mozFullScreen){
					canvas.mozRequestFullScreen();
				} else if (pointer.islocked()) {
					pointer.unlock();
				} else {
					document.mozCancelFullScreen();
				}
			}, false);*/
    
                    /*setTimeout(function() {
				if (!document.mozFullScreen){
					canvas.mozRequestFullScreen();
				} else if (pointer.islocked()) {
					pointer.unlock();
				} else {
					document.mozCancelFullScreen();
				}
			}, 2000);*/

                    document.addEventListener("mozfullscreenchange", function (e) {
                        if (document.mozFullScreen) {
                            pointer.lock(canvas);
                            is(pointer.islocked(), true, "Mouse should sdfdsfbe locked when DOM element is in the tree."); //if still fail, throw to inside the mozfullscreenchange event
                            document.mozCancelFullScreen();
                            /*if (pointer.islocked()) {
                                console.log("=== Mouse lock enabled but should not happen! ===");
                                //setTimeout("removeElement()", 2000);
                            }
                            /*else {
                                console.log("=== Mouse lock DISenabled! ===");
                            }
                            if (document.mozFullScreenElement === canvas) {
                                // Mouse should be lock if the element is in fullscreen mode
                                pointer.lock(canvas);
                                //is(pointer.islocked(), true, "Mouse should be locked when it gets a DOM element.");
                                pointer.unlock();
                                pointer.lock(canvas, function() {
                                    //is(pointer.islocked(), true, "Mouse should be locked when the success callback is fired.");
                                    lockTestDone = true;
                                    document.mozCancelFullScreen();
                                });
                            }
                            else {
                                // Mouse shoudn't lock if element isn't in fullscreen mode
                                pointer.lock(canvas);
                                //isnot(pointer.islocked(), true, "Mouse should only be locked if the locked element is in fullscreen mode");
                                unlockTestDone = true;
                                document.mozCancelFullScreen();
                            }*/
                        }
                        /*else {
                            if (lockTestDone === true && unlockTestDone === false) {
                                pointer.lock(canvas, function() {}, function() {
                                    //isnot(pointer.islocked(), true, "Mouse should be unlocked if a failure callback is fired.");
                                });
                                document.body.mozRequestFullScreen();
                            } else {
                                // Mouse should be unlocked if window is not in fullscreen mode
                                //isnot(pointer.islocked(), true, "Mouse should be unlocked if window is not in fullscreen mode");
                                //SimpleTest.finish();
                            }
                        }*/
                    }, false);
                    
                    /*function runTests() {
                        canvas.mozRequestFullScreen();
                    };*/
                    
                    // remove from DOM
                    /*function removeElement() {
                        console.log("=== Tag removed, mouselock disabled ===");
                        bodyTag.removeChild(canvas);
                    }*/
                    /*------------test end----------*/
                    SimpleTest.finish();
                });
            </script>
        </pre>
    </body>
</html>
